# speedup `date_trunc` (~7x faster)  in some cases