## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

- follow-up of #14593

## Rationale for this change

<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->

Follows the comment https://github.com/apache/datafusion/pull/14593#discussion_r1951334920 to implement an array version of `date_trunc`.

While implementing, I found that there is a lot of code handling boring calendar things, which is totally unrelated to some small granularities (less or equal to "hour"). So I also cut them off. Benchmark shows this simplification can bring 4x performance.

## What changes are included in this PR?

a faster implementation for `date_trunc` with granularities from "microsecond" to "hour"

```
date_trunc_minute_1000  time:   [16.600 µs 16.643 µs 16.684 µs]
Found 8 outliers among 100 measurements (8.00%)
  8 (8.00%) low mild

date_trunc_minute_1000  time:   [2.3474 µs 2.3519 µs 2.3579 µs]
                        change: [-85.946% -85.909% -85.870%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe
```

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

## Are these changes tested?

using existing unit tests

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->

## Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->